app/requirements.txt
flask
uwsgi
python-dotenv
flask-login
pyotp
bleach
passlib
pycryptodome
markdown
